perf: shrink EvalValue/LValue (float_type double; LValue field reorder)
The script float type changes from `long double` to `double`, and LValue's
`is_const` bool moves after `container_idx`. Together this shrinks the core
value objects:

  EvalValue: 48 -> 40 bytes   (long double forced 16-byte alignment, padding
                               EvalValue to 48; double is 8-aligned -> 40)
  LValue:    80 -> 56 bytes   (the smaller EvalValue, plus is_const packing
                               into the tail instead of wasting 7 bytes between
                               val and the container pointer)

LValue is the array element type, so the 30% drop cuts memory traffic for every
array op and every value copy. Measured speedups vs the prior build: array_append
1.31x, reverse 1.26x, sum 1.24x, foreach 1.22x/1.20x, concat 1.17x, subscript
1.13x, float_arith 1.13x, and math_builtins 2.27x (double libm is far faster
than the long double variants). Overall my/py geomean is now ~1.04x. Floats now
match Python's exactly (both 64-bit IEEE double); the only divergence left is
str() formatting (C %f fixed notation vs Python's shortest round-trip), which
the bench harness normalizes.

Tradeoff (accepted): double has ~15-16 significant digits vs long double's
~18-19, and an int beyond 2^53 mixed into a float loses exactness (long double's
64-bit mantissa held all 64-bit ints). Pure-int arithmetic is unchanged
(int_type stays intptr_t).

Mechanics: float parsing stold->stod (parser, float() builtin); the math
builtins switch from the `l`-suffixed libm functions to the double std::
overloads (selected by a function-pointer cast); fmodl/truncl/fabsl/roundl/powl
-> the double versions; str()'s snprintf format "%.*Lf" -> "%.*f" (this was also
a latent vararg bug: a double passed to a %Lf slot printed nan). size_type is
already uint32_t, so off/len needed no change. README / CLAUDE.md / bench/README
updated; the float-width divergence note is removed. 578 tests green, no
warnings, all 51 benchmark outputs match Python.

+++ src/defs.h
@@ -28,7 +28,14 @@ using std::make_shared;
 using std::make_unique;
 
 /* Custom type defs */
-typedef long double float_type; /* Note: replace %Lf if you change this! */
+/*
+ * The script float type. `double` (not `long double`): it keeps EvalValue small
+ * (long double's 16-byte alignment padded EvalValue from 40 to 48 bytes, which
+ * dominated array memory traffic and every value copy), matches Python's float
+ * exactly, and uses the much faster double libm. If you change this, update the
+ * printf/snprintf format strings (currently "%f"/"%.*f") and the math builtins.
+ */
+typedef double float_type;
 typedef intptr_t int_type;
 
 #ifndef _MSC_VER
